feat: v2 dataloader redesign — Time/TimeEncConfig, fast_val/test, dataset tests
v2 dataloader (batch.py / windowed.py / forecast.py / imputation.py):
- TSBatch gains x_time (Time dataclass) and x_time_feature (encoded Tensor)
- Time dataclass: .year .month .day .weekday .hour .minute .second (LongTensor)
- TimeEncConfig: bundles time_enc + freq; accepts custom_fn for bespoke encoders
- WindowConfig: remove include_raw/include_time/include_index/single_variate flags
  (all fields always included; None when unavailable)
- WindowConfig: add fast_val / fast_test flags (non-overlap stride for eval splits)
- Fix TimeEncConfig.encode: resolve string aliases through TimeEncoding before
  passing to time_features

Tests:
- tests/dataloader/test_v2_forecast.py: 39 tests covering shapes, Time components,
  index, stride, fast_val/test, column selection, TimeEncConfig, TSBatch.to()
- tests/dataset/test_forecast_datasets.py: 76 integration tests (marked slow)
  for ETTh1/h2/m1/m2, Weather, Traffic, Electricity, ILI, ExchangeRate

+++ torch_timeseries/experiments/Autoformer.py
@@ -109,7 +109,7 @@ class AutoformerUEAClassification(UEAClassificationExp, AutoformerParameters):
     model_type: str = "Autoformer"
 
     def _init_model(self):
-        self.label_len = self.windows/2
+        self.label_len = int(self.windows / 2)
         self.model = AutoformerClassModel(
             enc_in=self.dataset.num_features,
             dec_in=self.dataset.num_features,
@@ -168,7 +168,7 @@ class AutoformerAnomalyDetection(AnomalyDetectionExp, AutoformerParameters):
     model_type: str = "Autoformer"
 
     def _init_model(self):
-        self.label_len = self.windows/2
+        self.label_len = int(self.windows / 2)
         self.model = AnomalyModel(
             enc_in=self.dataset.num_features,
             dec_in=self.dataset.num_features,
@@ -222,7 +222,7 @@ class AutoformerImputation(ImputationExp, AutoformerParameters):
     model_type: str = "Autoformer"
 
     def _init_model(self):
-        self.label_len = self.windows/2
+        self.label_len = int(self.windows / 2)
         self.model = ImputationModel(
             enc_in=self.dataset.num_features,
             dec_in=self.dataset.num_features,
